# internal/pkg/docs/swagger.yaml
definitions:
  delivery.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  models.Attach:
    properties:
      id:
        type: string
      note_id:
        type: string
      path:
        type: string
    type: object
...
host: you-note.ru
info:
  contact: {}
  description: API for YouNote service
  title: YouNote API
  version: "1.0"
paths:
...
  /api/note/{id}/add_attach:
    post:
      consumes:
      - multipart/form-data
      description: Attach new file to note
      operationId: add-attach
      parameters:
      - description: note id
        in: path
        name: id
        required: true
        type: string
      - description: attach file
        in: formData
        name: attach
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: attach model
          schema:
            $ref: '#/definitions/models.Attach'
        "400":
          description: error
          schema:
            $ref: '#/definitions/delivery.ErrorResponse'
        "401":
          description: Unauthorized
        "413":
          description: Request Entity Too Large
      summary: Add attachment
      tags:
      - note
...
swagger: "2.0"